secure server actions

+++ src/app/[locale]/admin/locations/[id]/actions.ts
@@ -1,12 +1,21 @@
 "use server";
 
+import { authOptions } from "@/auth/auth";
 import { db } from "@/db/db";
 import { editLocationSchema, floors, locations } from "@/db/party/schema";
 import { eq } from "drizzle-orm";
+import { getServerSession } from "next-auth";
 import { revalidatePath } from "next/cache";
+import { redirect } from "next/navigation";
 import { z } from "zod";
 
 export async function editLocation(values: z.infer<typeof editLocationSchema>) {
+  const session = await getServerSession(authOptions);
+
+  if (!session) {
+    redirect("/api/auth/signin?error=Not%20Authenticated");
+  }
+
   try {
     const validated = editLocationSchema.parse(values);
 
@@ -28,8 +37,15 @@ const editFloorSchema = z.object({
 });
 
 export async function editFloor(value: z.infer<typeof editFloorSchema>) {
+  const session = await getServerSession(authOptions);
+
+  if (!session) {
+    redirect("/api/auth/signin?error=Not%20Authenticated");
+  }
+
   try {
     const validated = editFloorSchema.parse(value);
+
     await db.insert(floors).values(validated).onConflictDoNothing({
       target: floors.name,
     });
@@ -44,6 +60,12 @@ export async function editFloor(value: z.infer<typeof editFloorSchema>) {
 }
 
 export async function deleteFloor(id: number) {
+  const session = await getServerSession(authOptions);
+
+  if (!session) {
+    redirect("/api/auth/signin?error=Not%20Authenticated");
+  }
+
   try {
     await db.delete(floors).where(eq(floors.id, id));
